Improve and simplify user fetching.
